Actualizaciones y Nuevos ejercicios

+++ src/ej04.py
@@ -1,4 +1,4 @@
-celsius = int(input("Inserte una temperatura en celsius: "))
+celsius = float(input("Inserte una temperatura en celsius: "))
 
 farenheit = str(celsius * 9 / 5 + 32)
